new File can be created

+++ LevelEditor/LevelEditor/MainWindow.xaml.cs
@@ -25,13 +25,16 @@ namespace LevelEditor
     public partial class MainWindow : Window
     {
         ///<summary>directory info of File</summary>
-        public DirectoryInfo fileInfo = null;
+        public static DirectoryInfo fileInfo = null;
         ///<summary>filter for file</summary>
-        public static string filter = "Level File|*.lvl";
-        ///<summary>Level Info</summary>
+        public static string filter = "Level File|*.lvl|XML File|*.xml";
+        ///<summary>Current Level info</summary>
         public static Level level = new Level();
+        public static string path;
         ///<summary>true if something was changed</summary>
         public static bool changed = false;
+        /// <summary>Confirm to close Window</summary>
+        public static bool ConfirmClose;
 
         ///<summary>level manager private static</summary>
         private static LevelManager lvlManager = new LevelManager();
@@ -41,6 +44,8 @@ namespace LevelEditor
         private static Helper h = new Helper();
         ///<summary>All buttons</summary>
         private Button[,] allButtons = null;
+        ///<summary>Current Button</summary>
+        private static Button CurrentButton = null;
         ///<summary>All RowDefinitions</summary>
         private RowDefinition[] allRows = null;
         ///<summary>All ColumnDefinitions</summary>
@@ -49,12 +54,11 @@ namespace LevelEditor
         public MainWindow()
         {
             InitializeComponent();
-            TestFile();
 
             // Test
             UIElementCollection u = Grid_GridButtons.Children;
 
-            GenerateGridWithButtons(10, 10);
+            //GenerateGridWithButtons(10, 10);
             
             u = Grid_GridButtons.Children;
             var ItemInFirstRow = u.Cast<UIElement>().Where(i => Grid.GetRow(i) == 0);
@@ -142,7 +146,60 @@ namespace LevelEditor
         /// <param name="e">The <see cref="RoutedEventArgs"/> instance containing the event data.</param>
         private void Button_New_Click(object sender, RoutedEventArgs e)
         {
+            bool? result;
+            bool changedTmp = WasChanged(out result);
 
+            if (changedTmp == true)
+            {
+                switch (result)
+                {
+                    // save
+                    case true:
+                        // if file already exist
+                        if (fileInfo == null)
+                        {
+                            h.SaveFile(out bool? work);
+
+                            // if file was not saved returrn
+                            if (work != true)
+                                return;
+                        }
+                        break;
+                    // jump directly to creating new file
+                    case false:
+                        break;
+                    // Cancel creating file
+                    case null:
+                        return;
+                }
+            }
+
+            // prepare new level
+            NewLevel prepare = new NewLevel();
+            prepare.ShowDialog();
+            List<Tile> tilesList = new List<Tile>();
+            List<Layer> layerList = new List<Layer>();
+            Layer layer = new Layer();
+            layer.ZOrder = 0;
+            layerList.Add(layer);
+
+            // return if levelname is null
+            if (prepare.levelName == null)
+                return;
+
+            // delete old grid and buttons
+            DeleteAllButtons();
+            DeleteGrid();
+
+            // generate new level
+            level = new Level();
+            level.Name = prepare.levelName;
+            level.SizeX = prepare.xSize;
+            level.SizeY = prepare.ySize;
+            level.Layer = layerList;
+            GenerateGridWithButtons(level.SizeY, level.SizeX);
+            changed = true;
+            path = null;
         }
 
         /// <summary>
@@ -169,9 +226,8 @@ namespace LevelEditor
                 }
 
                 // set level Path
-                level.Path = file.FileName;
-                fileInfo = new DirectoryInfo(level.Path);
-                SetStatus(Label_StatusbarOne, "Loading File...", true, level.Path);
+                path = file.FileName;
+                SetStatus(Label_StatusbarOne, "Loading File...", true, path);
             }
             // User canceled window
             else
@@ -181,8 +237,14 @@ namespace LevelEditor
             }
 
             // loading File
-            LvlManager.LoadLevel(level.Path);
+            level = LvlManager.LoadLevel(path);
 
+            // delete old grid and buttons
+            DeleteEverything();
+
+            // set grid and buttons
+            GenerateGridWithButtons(level.SizeY, level.SizeX);
+            
             // loading complete. set statusbar
             SetStatus(Label_StatusbarOne, "Loading Complete", true);
         }
@@ -204,7 +266,13 @@ namespace LevelEditor
         /// <param name="e">The <see cref="RoutedEventArgs"/> instance containing the event data.</param>
         private void Button_Save_Click(object sender, RoutedEventArgs e)
         {
-            lvlManager.SaveLevel(level.Path + "ll", level);
+            if (path == null)
+                Button_SaveAs_Click(sender, e);
+            else
+            {
+                lvlManager.SaveLevel(path, level);
+                ConfirmClose = true;
+            }
         }
 
         /// <summary>
@@ -214,7 +282,11 @@ namespace LevelEditor
         /// <param name="e">The <see cref="RoutedEventArgs"/> instance containing the event data.</param>
         private void Button_SaveAs_Click(object sender, RoutedEventArgs e)
         {
-
+            h.SaveFile(out bool? succellful);
+            if (succellful == true)
+                ConfirmClose = true;
+            else
+                ConfirmClose = false;
         }
 
 
@@ -233,7 +305,66 @@ namespace LevelEditor
         #endregion
         #endregion
 
-        #region Tiles Buttons
+        #region Properties Window
+
+        #region Name
+        private void TextBox_SpriteID_TextChanged(object sender, TextChangedEventArgs e)
+        {
+            List<Layer> newLayer = level.Layer;
+            
+            level.Name = TextBox_SpriteID.Text;
+        }
+        #endregion
+
+        #region PosX
+        private void TextBox_PosX_TextChanged(object sender, TextChangedEventArgs e)
+        {
+
+        }
+        #endregion
+
+        #region PosY
+        private void TextBox_PosY_TextChanged(object sender, TextChangedEventArgs e)
+        {
+
+        }
+        #endregion
+
+        #region Comment
+        private void TextBox_Comment_TextChanged(object sender, TextChangedEventArgs e)
+        {
+
+        }
+        #endregion
+
+        #region Collision
+
+        private void CheckBox_Collision_Checked(object sender, RoutedEventArgs e)
+        {
+            // return if no button was chosen
+            if (CurrentButton == null)
+                return;
+
+            string Tag = CurrentButton.Tag.ToString();
+        }
+
+        private void CheckBox_Collision_Unchecked(object sender, RoutedEventArgs e)
+        {
+            // return if no button was chosen
+            if (CurrentButton == null)
+                return;
+
+            string Tag = CurrentButton.Tag.ToString();
+
+        }
+        #endregion
+
+        #region Tag
+        private void TextBox_Tag_TextChanged(object sender, TextChangedEventArgs e)
+        {
+
+        }
+        #endregion
 
         #endregion
 
@@ -245,29 +376,36 @@ namespace LevelEditor
         /// <param name="e">The <see cref="RoutedEventArgs"/> instance containing the event data.</param>
         private void bttn_Click(object sender, RoutedEventArgs e)
         {
-            Button bttn = (Button)sender;
-            string tag = bttn.Tag.ToString();
-            bttn.Background = Brushes.Aqua;
-
+            CurrentButton = (Button)sender;
+            changed = true;
         }
         #endregion
-        private void Button_DevOnly1_Click(object sender, RoutedEventArgs e)
+        private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
         {
-            GenerateGridWithButtons(20, 20);
-        }
+            ConfirmClose = false;
+            bool changed = WasChanged(out bool? result);
+            if (changed == false)
+                ConfirmClose = true;
 
-        private void Button_DevOnly2_Click(object sender, RoutedEventArgs e)
-        {
-            DeleteAllButtons();
-            DeleteGrid();
-        }
-        // --------------------------------- //
+            else if (result == true)
+                Button_Save_Click(sender, new RoutedEventArgs());
+            else if (result == false)
+                ConfirmClose = true;
+            else
+            {
+                e.Cancel = ConfirmClose;
+            }
 
+            e.Cancel = !ConfirmClose;
+        }
+        // ----------------------------------------------------------------------------------------------------- //
+
+        #region Create Grid and Buttons
         /// <summary>
         /// Generates the grid with buttons.
         /// </summary>
-        /// <param name="_rowDef">height of row</param>
-        /// <param name="_columnDef">width of column</param>
+        /// <param name="_rowDef">height of row (Y Size)</param>
+        /// <param name="_columnDef">width of column (X Size)</param>
         private void GenerateGridWithButtons(int _rowDef, int _columnDef)
         {
             // generate All Buttons array
@@ -327,9 +465,11 @@ namespace LevelEditor
             bttn.Tag = _row + "|" + _col;
             bttn.Content = _buttonContent;
             bttn.Click += bttn_Click;
+            bttn.Focusable = false;
 
             return bttn;
         }
+        #endregion
 
         #region Show Button Function
         /// <summary>
@@ -371,6 +511,8 @@ namespace LevelEditor
         }
         #endregion
 
+        #region Delete Functions
+
         /// <summary>
         /// Deletes all buttons.
         /// </summary>
@@ -380,17 +522,75 @@ namespace LevelEditor
         }
 
         /// <summary>
-        /// Deletes the grid.
+        /// Delete complete grid (WITHOUT BUTTONS).
         /// </summary>
         private void DeleteGrid()
         {
+            // Delete Rows
+            DeleteRow();
+            // Delete Columns
+            DeleteColumn();
+        }
+
+        /// <summary>
+        /// Delete row
+        /// </summary>
+        private void DeleteRow()
+        {
+            // check size of row definition
+            if (Grid_GridButtons.RowDefinitions.Count == 0)
+                return;
+
             // delete RowDefinitions
             Grid_GridButtons.RowDefinitions.RemoveRange(0, Grid_GridButtons.RowDefinitions.Count);
+        }
+
+        /// <summary>
+        /// Delete column
+        /// </summary>
+        private void DeleteColumn()
+        {
+            // check size of row definition
+            if (Grid_GridButtons.ColumnDefinitions.Count == 0)
+                return;
 
             // delete ColumnDefinitions
             Grid_GridButtons.ColumnDefinitions.RemoveRange(0, Grid_GridButtons.ColumnDefinitions.Count);
         }
 
+        /// <summary>
+        /// Delete Grid and Buttons
+        /// </summary>
+        private void DeleteEverything()
+        {
+            DeleteAllButtons();
+            DeleteGrid();
+        }
+        #endregion
 
+        /// <summary>
+        /// check if file was changed and ask to save if file was changed.
+        /// </summary>
+        /// <param name="MsgBoxResult">what User pressed</param>
+        /// <returns>true: file was changed</returns>
+        public static bool WasChanged(out bool? MsgBoxResult)
+        {
+            MsgBoxResult = null;
+            if(changed == false)
+            {
+                return false;
+            }
+
+            MessageBoxResult result = MessageBox.Show("Would you like to save?", "Unsaved Changes", MessageBoxButton.YesNoCancel);
+
+            if (result == System.Windows.MessageBoxResult.Yes)
+                MsgBoxResult = true;
+            else if (result == System.Windows.MessageBoxResult.No)
+                MsgBoxResult = false;
+            else
+                MsgBoxResult = null;
+
+            return true;
+        }
     }
 }
